fix begin at 0xa0800200    fix length problem in creatimage  add page swap

+++ tools/createimage.c
@@ -8,7 +8,7 @@
 
 void write_bootblock(FILE *image, FILE *bbfile, Elf32_Phdr *Phdr);
 Elf32_Phdr *read_exec_file(FILE *opfile);
-uint8_t count_kernel_sectors(Elf32_Phdr *Phdr);
+int count_kernel_sectors(Elf32_Phdr *Phdr);
 void extent_opt(Elf32_Phdr *Phdr_bb, Elf32_Phdr *Phdr_k, int kernelsz);
 
 void write_filesegement_image(FILE *image, FILE *file, Elf32_Phdr *phdr)
@@ -76,7 +76,7 @@ Elf32_Phdr *read_exec_file(FILE *opfile)
     return Pro_header;
 }
 
-uint8_t count_kernel_sectors(Elf32_Phdr *Phdr)
+int count_kernel_sectors(Elf32_Phdr *Phdr)
 {
 	return  (Phdr->p_memsz - 1)/ 512 + 1;
 }
@@ -101,7 +101,7 @@ void write_kernel(FILE *image, FILE *knfile, Elf32_Phdr *Phdr, int kernelsz)
 	write_filesegement_image(image, knfile, Phdr);
 }
 //在image上直接更改第一扇区的最后的值为kernel扇区数
-void record_kernel_sectors(FILE *image, uint8_t kernelsz)
+void record_kernel_sectors(FILE *image, int kernelsz)
 {
 	int buffer=kernelsz;
 	fseek(image,508,SEEK_SET);
